app/services/oauth.py
import httpx # type: ignore
from typing import Optional
from app.utils.logger import setup_logger

logger = setup_logger("oauth_service", "oauth.log")

class OAuthService:
    async def verify_google_token(self, token: str) -> Optional[dict]:
        try:
            async with httpx.AsyncClient() as client:
                response = await client.get(
                    f"https://oauth2.googleapis.com/tokeninfo?id_token={token}"
                )
                if response.status_code == 200:
                    return response.json()
                return None
        except Exception as e:
            logger.error(f"Google token verification failed: {str(e)}")
            return None

    async def verify_apple_token(self, token: str) -> Optional[dict]:
        try:
            # Implement Apple token verification
            # Refer to Apple's documentation for proper implementation
            pass
        except Exception as e:
            logger.error(f"Apple token verification failed: {str(e)}")
            return None

oauth_service = OAuthService()